requirements.txt
flask==2.3.3
flask-cors>=4.0.0
pandas>=2.0.0
pyodbc>=4.0.34
ollama>=0.1.5
chromadb>=0.4.18
sentence-transformers>=2.2.2
vanna>=0.3.0
plotly>=5.14.0
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.28.0
